requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.11.11
python-dotenv==1.0.0
aiohttp==3.9.5
aiosqlite==0.20.0
psutil==5.9.0